requirements.txt
streamlit
pandas
numpy
sentence-transformers
scikit-learn
google-generativeai